nit: Move NetworkTask to its own file & reorder NetworkData functions
